[#358] canceled meeting email

+++ backend/EasyMeets.Core/EasyMeets.Core.BLL/Interfaces/IMeetingService.cs
@@ -6,7 +6,7 @@ namespace EasyMeets.Core.BLL.Interfaces
     public interface IMeetingService
     {
         Task<List<MeetingSlotDTO>> GetMeetingsAsync(MeetingMemberRequestDto meetingMemberRequestDto);
-        Task<List<UserMeetingDTO>> GetAllMembers(int id);
+        Task<List<UserMeetingDTO>> GetAllMembers(long id);
         Task<List<OrderedMeetingTimesDto>> GetOrderedMeetingTimesAsync(long slotId);
         Task<SaveMeetingDto> CreateMeeting(SaveMeetingDto meetingDto);
         Task<long> CreateExternalAttendeeMeeting(ExternalAttendeeMeetingDto meetingDto);
